Add metrics e2e test
An e2e test has been added that uses the new mock Segment server to
verify that the `Close Tx Via Notification Close` event is being sent
correctly when the notification window closes.

Using the mock Segment server requires a special build with this mock
Segment server hostname embedded, so a distinct job for building and
running this test was required. As such, it was left out of the
`run-all.sh` script.